pyproject.toml
[project]
name = "my-blog-agent"
version = "0.1.0"
description = "Multi-agent blog writer using ADK + OpenAI GPT-4o mini"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
  "google-adk",
  "python-dotenv",
  # OpenAI support
  "openai",
  # Optional wrapper (recommended)
  "litellm",
  # RAG (we'll use later)
  "llama-index",
  # optional
  "tqdm",
  "langchain-community>=0.4.1",
  "duckduckgo-search>=8.1.1",
  "ddgs>=9.13.1",
]

[tool.uv]
